Add validation losses to training loop
Add validation losses and use use to determine when to stop training.
Using training losses for this is not a good idea because it can lead
to overfitting and it would look like the loss would just keep
improving.
Add training and validation losses to a plot.
Add more verbose printouts during training.
Rename EXPERIMENT_EVENT_HORIZON to PREDICTION_WINDOW.