Add support for label weights in the loss function
Originally I thought that this would work with BCELoss(), but it
requires BCELossWithLogits() instead. This is because the former
expects the input to be a probability, while the latter expects
the input to be a logit.

To use BCELossWithLogits(), we need to remove the final sigmoid
activation from the model. This is now selectable, depending on the
loss function.